Optimize DQN throughput via FP16 tensor cores and FP16 buffer storage
Two targeted optimizations reduce learning-phase latency and memory:

1. FP16 mixed precision in _learn_phase: cast params+obs to float16
   before each forward pass (target and online network), cast output
   back to float32 for stable loss. Activates Turing tensor cores
   (turing_fp16_s1688gemm) replacing FP32 CUDA GEMMs (volta_sgemm).
   GEMM time drops from ~728ms to ~117ms per scan step.

2. FP16 buffer storage: cast observations to float16 before buffer.add().
   MinAtar observations are 0/1 booleans, exactly representable in FP16.
   Halves buffer memory (639→324 MiB) and buffer.sample() gather
   bandwidth (loop_slice_fusion: 301→149ms per scan step).

Combined result: +80-85% throughput (90k→163-167k items/s steady-state),
-49% GPU memory, reward quality preserved (identical loss/returns curves).
Profiler evidence in artifacts/profiles/; benchmark runs in artifacts/benchmarks/.

+++ benchmarks/purejaxrl/dqn.py
@@ -89,7 +89,9 @@ def make_train(config):
         _action = basic_env.action_space().sample(rng)
         _, _env_state = env.reset(rng, env_params)
         _obs, _, _reward, _done, _ = env.step(rng, _env_state, _action, env_params)
-        _timestep = TimeStep(obs=_obs, action=_action, reward=_reward, done=_done)
+        # Store observations as FP16 in buffer: MinAtar obs are 0/1 booleans
+        # (exact in FP16), halving buffer memory and sampling bandwidth.
+        _timestep = TimeStep(obs=_obs.astype(jnp.float16), action=_action, reward=_reward, done=_done)
         buffer_state = buffer.init(_timestep)
 
         # INIT NETWORK AND OPTIMIZER
@@ -157,8 +159,8 @@ def make_train(config):
                 timesteps=train_state.timesteps + config["NUM_ENVS"]
             )  # update timesteps count
 
-            # BUFFER UPDATE
-            timestep = TimeStep(obs=last_obs, action=action, reward=reward, done=done)
+            # BUFFER UPDATE: cast obs to FP16 to halve buffer memory and gather bandwidth
+            timestep = TimeStep(obs=last_obs.astype(jnp.float16), action=action, reward=reward, done=done)
             buffer_state = buffer.add(buffer_state, timestep)
 
             # NETWORKS UPDATE
